+++ .vim/coc-settings.json
@@ -1,0 +1,4 @@
+{
+  "tsserver.useLocalTsdk": true,
+  "tsserver.tsdk": "${workspaceFolder}/.yarn/sdks/typescript/lib"
+}
